auto create & delete secret

+++ e2e/stop_gcp_enclave.sh
@@ -1,10 +1,8 @@
 #!/usr/bin/env bash
 set -ex
 
-PROJECT=uid2-test
-
-if [ -z "$GCP_INSTANCE_NAME" ]; then
-  echo "GCP_INSTANCE_NAME can not be empty"
+if [ -z "$GCP_PROJECT" ]; then
+  echo "GCP_PROJECT can not be empty"
   exit 1
 fi
 
@@ -13,9 +11,18 @@ if [ -z "$SERVICE_ACCOUNT" ]; then
   exit 1
 fi
 
-gcloud config set project $PROJECT
+if [ -z "$GCP_INSTANCE_NAME" ]; then
+  echo "GCP_INSTANCE_NAME can not be empty"
+  exit 1
+fi
+
+OPERATOR_KEY_SECRET_NAME=$GCP_INSTANCE_NAME
+
+gcloud config set project $GCP_PROJECT
 
 gcloud config set compute/zone asia-southeast1-a
 
 gcloud compute instances delete $GCP_INSTANCE_NAME \
     --quiet
+
+gcloud secrets delete $OPERATOR_KEY_SECRET_NAME
